feat: make easy tests pass
