+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/limits/L1BlockSize.java
@@ -15,6 +15,8 @@
 
 package net.consensys.linea.zktracer.module.limits;
 
+import static net.consensys.linea.zktracer.module.ModuleName.BLOCK_L1_SIZE;
+
 import java.util.List;
 
 import lombok.Getter;
@@ -24,11 +26,12 @@ import net.consensys.linea.zktracer.Trace;
 import net.consensys.linea.zktracer.container.module.IncrementingModule;
 import net.consensys.linea.zktracer.container.module.Module;
 import net.consensys.linea.zktracer.container.stacked.CountOnlyOperation;
-import org.apache.tuweni.bytes.Bytes;
+import net.consensys.linea.zktracer.types.TransactionProcessingMetadata;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Hash;
 import org.hyperledger.besu.datatypes.Transaction;
 import org.hyperledger.besu.evm.log.Log;
+import org.hyperledger.besu.evm.log.LogTopic;
 import org.hyperledger.besu.evm.worldstate.WorldView;
 import org.hyperledger.besu.plugin.data.ProcessableBlockHeader;
 
@@ -37,15 +40,14 @@ import org.hyperledger.besu.plugin.data.ProcessableBlockHeader;
 @RequiredArgsConstructor
 public class L1BlockSize implements Module {
 
+  private final BlockTransactions blockTransactions;
+  private final Keccak keccak;
   private final IncrementingModule l2l1Logs;
   private final Address l2l1Address;
-  private final Bytes l2l1Topic;
+  private final LogTopic l2l1Topic;
 
   private final short TIMESTAMP_BYTESIZE = 32 / 8;
   private final short NB_TX_IN_BLOCK_BYTESIZE = 16 / 8;
-
-  /** A simple counter for the number of transactions */
-  private final CountOnlyOperation numberOfTransactions = new CountOnlyOperation();
 
   /** The byte size of the RLP-encoded transaction of the conflation */
   private final CountOnlyOperation sizesRlpEncodedTxs = new CountOnlyOperation();
@@ -58,19 +60,17 @@ public class L1BlockSize implements Module {
 
   @Override
   public String moduleKey() {
-    return "BLOCK_L1_SIZE";
+    return BLOCK_L1_SIZE.toString();
   }
 
   @Override
   public void commitTransactionBundle() {
-    numberOfTransactions.commitTransactionBundle();
     sizesRlpEncodedTxs.commitTransactionBundle();
     l2l1LogSizes.commitTransactionBundle();
   }
 
   @Override
   public void popTransactionBundle() {
-    numberOfTransactions.popTransactionBundle();
     sizesRlpEncodedTxs.popTransactionBundle();
     l2l1LogSizes.popTransactionBundle();
   }
@@ -82,7 +82,7 @@ public class L1BlockSize implements Module {
 
         // Calculates the data size related to the abi encoding of the list of the
         // from addresses. The field is a simple array of bytes20.
-        + numberOfTransactions.lineCount() * Address.SIZE
+        + blockTransactions.lineCount() * Address.SIZE
 
         // Calculates the data size related to the block
         + nbBlock * (TIMESTAMP_BYTESIZE + Hash.SIZE + NB_TX_IN_BLOCK_BYTESIZE);
@@ -95,12 +95,15 @@ public class L1BlockSize implements Module {
 
   @Override
   public List<Trace.ColumnHeader> columnHeaders(Trace trace) {
-    throw new IllegalStateException("should never be called");
+    throw new IllegalStateException("non-tracing module");
+  }
+
+  @Override
+  public void traceEndTx(TransactionProcessingMetadata tx) {
+    traceEndTx(tx.getBesuTransaction(), tx.getLogs());
   }
 
   public void traceEndTx(Transaction tx, List<Log> logs) {
-    numberOfTransactions.add(1);
-
     for (Log log : logs) {
       if (isL2L1Log(log)) {
         l2l1LogSizes.add(log.getData().size());
@@ -117,6 +120,12 @@ public class L1BlockSize implements Module {
     // incurred by the top-level array, hence the +1.
     final int txDataSize = tx.encoded().size();
     sizesRlpEncodedTxs.add(txDataSize);
+    // Counts the number of Keccak from tx RLPs, used both for both the signature verification and
+    // the public input computation.
+    keccak.updateTally(txDataSize);
+    // Counts the number of Keccak from tx RLPs preimage (RLP of the transaction wo the signature)
+    final int txPreimageByteSize = tx.encodedPreimage().size();
+    keccak.updateTally(txPreimageByteSize);
   }
 
   @Override
